refactor: welcome・recordsページをページエントリーポイント方式に移行
