+++ resources/js/pages/records-index.js
@@ -1,0 +1,9 @@
+import RecordsFilterView from '../features/records/filter-view.js';
+
+/**
+ * Records Index ページの初期化
+ */
+document.addEventListener('DOMContentLoaded', () => {
+    // Records フィルター表示機能を初期化
+    new RecordsFilterView();
+});
