+++ resources/js/pages/welcome.js
